chart-pattern-intel/backend/requirements.txt
# Backend dependencies for Chart Pattern Intelligence
fastapi
uvicorn
yfinance
pandas
python-dotenv
pydantic
google-generativeai
requests
